porto: let TClient instance be created without TEpollSource backing
Behaviour of some properties depends on the client context
along with the container state itself.

E.g. when we are restoring values from the storage, we actually
have no client, so this feature will let us create "fake" root
without any client associated, which can be used to restore values
in the "forced" fashion.

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>
Signed-off-by: Konstantin Khlebnikov <khlebnikov@yandex-team.ru>