porto: add set state tracking for reworked knobs
Container initialization depends on if needed knob
has user-defined value or default, e.g. path to stdout/stderr.
This is also important on the restore phase for the already running
container, so we need to restore info on which knobs was set.

Use simple bit mask inside the container to store knobs state,
save only knobs which set by the user and restore knobs in set state.

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>