porto: fix hierarchy memory guarantee accounting
Under some circumstances it is possible to create
configuration with overcommitment on memory guarantees
due to bug in guarantees accounting which lead to skipping
container's children guarantees.

Fix by defining memory guarantee of a hierarchy as
the maximum between the sum of children container guarantees
and the guarantee of root.

Also rework the memory guarantee container property handler code.

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>