porto: implement separate interface for knob saving/restoring
There are knobs, which are read-only from user perspective,
but it is appropriate to serialize their values (e.g. start or
death time), thus we need to set read-only knobs internally.

Since we are using knobs both for client requests processing
and value serialization, let's implement separate interface
to make code clearer.

This patch is a part of the property reworking patchset.

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>