porto: save container state after each state change.
For reworked properties TContainer::Save()
should be called explicitly where needed.

The current state of container should be saved, so
we need to dump container context after each container
state change. The other case is since we have post-mortem
persistent properties and data, then we need to dump them
too.

For pause and resume Save() is called for each child.
In the kill handler we don't need to save state,
we just send event to the container, which will
be manifested as exit later.

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>