porto: fix broken container stop semantics
Previously it was declared that porto stop operation
is performed by sending SIGTERM to all processes in
the container, waiting 30 second to give applications
ability to properly clean everything up and if any
process stayed alive after the timout, it should be
killed with SIGKILL.

Unfortunately, this contract was broken - the actual
timeout was strictly 1 second between SIGTERM and SIGKILL.

Add optional timeout parameter for Stop RPC to control
desired duration of the graceful shutdown. If container has
children, then the timeout value is the total time limit
for the whole hierarchy to gracefully stop.

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>