porto: do not respawn container if parent is not running
Respawn only if the parent container is still alive.

This patch fixes bug with portod reaching inconsistent
state with respawned container to be in the running state
while its parent is dead.

The container hierarchy respawn was broken even before this
change, may be it should be fixed later.

This case can be reproduced by setting respawn for child
and terminating parent container from inside (e.g. by kill -9).

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>
Signed-off-by: Konstantin Khlebnikov <khlebnikov@yandex-team.ru>