volume: fix for segfault on broken volume restore
If porto was forced interrupted during volume operation
(e.g. volume export or creation), or volume kev-value
node became broken, portod segfaults on restart.

The problem is that in this case the TVolume::Restore()
fails and the TVolume::Destroy() is called, and at this
point the designated container is not associated with
volume being restored.

However, in Destroy() portod does not check the case when
the container didn't have the link to the volume.

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>