container: fix stopping hierarchies with paused children
Portod incorrectly handles Stop command for non-paused
containers with paused children. This leads to situation
that paused children are not terminated correctly,
which jeopardizes cgroup cleanup and container disposal.

Fix it with 2-pass hierarchy stop process:
on the first pass terminate and thaw children if necessary,
on the second perform cleanup.

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>
Signed-off-by: Konstantin Khlebnikov <khlebnikov@yandex-team.ru>