container: do not treat ro container place on start as tragedy #PORTO-238
Sudden read-only remount of a partition is definetely
critical event but it is not portod fault after all.

So there is no need to shout about it on each container
start try, log message and failed start increment is enough.

Fix inconsistent container state in case of resource
preparation fail.

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>
Signed-off-by: Konstantin Khlebnikov <khlebnikov@yandex-team.ru>