cgroup: add systemd cgroup type
Cgroup with type "name=systemd" is required to be
mounted in order to execute systemd in container.

This cgroup is writeable inside the container,
thus should be cleared by porto on remove.

The whole hierarchy is mounted ro inside the chroot,
but the container subtree is writeable.

Systemd takes cgroup prefix from /proc/self/cgroups and
since the value is not virtualized, so the systemd
should select the correct prefix inside mounted systemd
cg tree.

Signed-off-by: Maxim Samoylov <max7255@yandex-team.ru>
Signed-off-by: Konstantin Khlebnikov <khlebnikov@yandex-team.ru> (rewrite, add autodetect)